+++ tests/api/conftest.py
@@ -1,0 +1,45 @@
+import pytest
+from fastapi.testclient import TestClient
+from pytest_bdd import given
+
+from app.database import get_db
+from app.main import app
+from tests.config import HEALTH_ENDPOINT
+
+# The "db" fixture is automatically available here from the common conftest.py
+
+
+# Provide a TestClient fixture that overrides the database dependency
+@pytest.fixture(scope="session")
+def client(db):
+    def override_get_db():
+        try:
+            yield db
+        finally:
+            db.close()
+
+    app.dependency_overrides[get_db] = override_get_db
+    client = TestClient(app)
+    yield client
+    app.dependency_overrides.clear()
+
+
+# BDD step to check that the API is running
+@given("the API is running")
+def api_is_running(client):
+    response = client.get(HEALTH_ENDPOINT)
+    assert response.status_code == 200
+    health_json = response.json()
+    assert health_json["status"] == "healthy"
+    assert health_json["components"]["api"] == "healthy"
+    assert health_json["components"]["ui"] == "healthy"
+
+
+@pytest.fixture
+def mark_token(client):
+    def _mark_token(token, is_stable):
+        payload = {"token": token, "is_stable": is_stable}
+        response = client.post("/api/tokens", json=payload)
+        assert response.status_code == 200
+
+    return _mark_token
